# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: api
  version: 1.0.0
  description: EngineeringOS - Autonomous AI Engineering Platform API

# PR-02: Global security default — all /api/* routes require a valid Clerk
# session unless an operation explicitly overrides with `security: []`.
# This ensures new routes are protected by default without having to remember
# to add a per-operation security declaration.
security:
  - ClerkSession: []
paths:
...
  /api/tasks/{taskId}/execute:
    post:
      operationId: executeTask
      summary: Execute a task through the AI pipeline
      tags: [Tasks]
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Task execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
  /api/ai/tasks/{taskId}/execute:
    post:
      operationId: aiExecuteTask
      summary: Execute a task via AI agent
      tags: [AI]
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Task executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  securitySchemes:
    ClerkSession:
      type: apiKey
      in: cookie
      name: __session
      description: >
        Clerk session cookie. All /api/* routes (except /api/healthz) require a
        valid Clerk session. The cookie is set automatically by the Clerk frontend
        SDK after sign-in — no manual header injection is needed.
  schemas:
...
    TaskStatus:
      type: string
      enum: [pending, queued, running, verifying, completed, failed, cancelled]
    TaskPriority:
      type: string
      enum: [p0, p1, p2, p3]
...
    Task:
      type: object
      required: [id, projectId, title, status, priority, createdAt, updatedAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        ruleId:
          type: string
        workflowId:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        relatedFiles:
          type: array
          items:
            type: string
        dependsOn:
          type: array
          items:
            type: string
        retryCount:
          type: integer
        maxRetries:
          type: integer
        phase:
          type: string
        prompt:
          type: string
        agentResponse:
          type: string
        verificationResult:
          $ref: "#/components/schemas/VerificationResult"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
    VerificationResult:
      type: object
      required: [passed, steps]
      properties:
        passed:
          type: boolean
        steps:
          type: array
          items:
            type: object
            required: [name, passed]
            properties:
              name:
                type: string
              passed:
                type: boolean
              output:
                type: string

    # ─── Rule ──────────────────────────────────────────────────────────────────
...